# docs/kaayko-paddling-api-swagger.yaml
openapi: 3.0.3
info:
  title: Kaayko Paddling Intelligence APIs
  version: "2.1.0"
  description: |
    # Kaayko Paddling Intelligence APIs v2.1.0
    
    **Enterprise-grade paddling condition intelligence powered by machine learning and real-time marine data integration**
    
    ---
    
    ## 🎯 Executive Summary
    
    The Kaayko Paddling Intelligence APIs provide ultra-fast, machine learning-powered paddling condition assessments with comprehensive safety analysis. Our platform processes real-time weather, marine, and environmental data to deliver actionable insights for paddlers across 18+ premium locations.
    
    ### Core Capabilities
    
    | API Endpoint | Purpose | Response Time | Key Features |
    |--------------|---------|---------------|--------------|
    | `paddlingOut` | Location Discovery | ~100ms | Comprehensive location database with amenities |
    | `paddleScore` | Real-time Conditions | ~250ms | Current conditions with ML-powered safety ratings |
    | `fastForecast` | 3-Day Predictions | <4s fresh / <1s cached | Hourly forecasts with penalty breakdown |
    | `forecast` | Premium Analytics | Variable | Advanced analytics for professional use |
    
    ### 🚀 Version 2.1.0 Enhancements
    
    - **🎯 Precision Rating System**: 0.5-increment ratings (1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0)
    - **🌊 Marine Data Integration**: Real-time wave height and water temperature analysis
...
    
    ## 🧠 Machine Learning Intelligence
    
...
    ```
...
    
...
servers:
...
security:
  - ApiKeyAuth: []
...
paths:
  /paddlingOut:
    get:
      tags:
        - paddlingOut
      operationId: getPaddlingLocations
      summary: Retrieve all paddling locations
      description: |
        **Comprehensive paddling location database with amenities and coordinates**
        
        This endpoint provides access to our curated database of premium paddling locations across North America. Each location includes precise coordinates, detailed amenities information, and multimedia resources.
        
        ### Key Features
        - ✅ **18+ Premium Locations** - Hand-selected paddling destinations
        - 🎯 **Precise Coordinates** - GPS coordinates for navigation systems
        - 🅿️ **Amenity Details** - Parking, restrooms, and facility information
        - 📸 **Rich Media** - Photos and instructional videos
        - 🏞️ **Location Context** - Difficulty ratings and paddling tips
        
        ### Use Case
        Use the coordinates from this endpoint for all subsequent API calls (`paddleScore`, `fastForecast`, `forecast`) to ensure marine data accuracy and optimal predictions.
        
        ### Response Time
        **~100ms** - Optimized for mobile and web applications
        
      responses:
        '200':
          description: Successfully retrieved paddling locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaddlingLocation'
              examples:
                complete_response:
                  summary: Complete locations response
                  description: Array of paddling locations with full details
                  value:
                    - id: "trinity_river"
                      lakeName: "Trinity River"
                      title: "Trinity River"
                      subtitle: "Irving, Texas"
                      text: "Solitude, birds and a good workout with consistent flow"
                      location:
                        latitude: 32.881187
                        longitude: -96.929937
                      parkingAvl: "Y"
                      restroomsAvl: "Y"
                      youtubeURL: "https://youtube.com/watch?v=trinity_river_guide"
                      imgSrc: 
                        - "https://firebasestorage.googleapis.com/v0/b/kaaykostore.appspot.com/o/images%2Fpaddling_out%2Ftrinity_river_1.jpg?alt=media"
                        - "https://firebasestorage.googleapis.com/v0/b/kaaykostore.appspot.com/o/images%2Fpaddling_out%2Ftrinity_river_2.jpg?alt=media"
                    - id: "cottonwood_lake"
                      lakeName: "Cottonwood Lake"
                      title: "Cottonwood Lake"
                      subtitle: "Buena Vista, Colorado"
                      text: "High altitude alpine paddling with stunning mountain views"
                      location:
                        latitude: 38.781063
                        longitude: -106.277812
                      parkingAvl: "Y"
                      restroomsAvl: "N"
                      youtubeURL: ""
                      imgSrc: 
                        - "https://firebasestorage.googleapis.com/v0/b/kaaykostore.appspot.com/o/images%2Fpaddling_out%2Fcottonwood_lake_1.jpg?alt=media"
                    - id: "lake_crescent"
                      lakeName: "Lake Crescent"
                      title: "Lake Crescent"
                      subtitle: "Olympic National Park, Washington"
                      text: "Crystal clear glacial lake surrounded by old-growth forest"
                      location:
                        latitude: 48.052813
                        longitude: -123.870438
                      parkingAvl: "Y"
                      restroomsAvl: "Y"
                      youtubeURL: ""
                      imgSrc:
                        - "https://firebasestorage.googleapis.com/v0/b/kaaykostore.appspot.com/o/images%2Fpaddling_out%2Flake_crescent_1.jpg?alt=media"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Server error"
                details: "Database connection failed"
                timestamp: "2025-08-23T10:30:00Z"
...
  /paddleScore:
    get:
      tags:
        - paddleScore
      operationId: getCurrentPaddleScore
      summary: Get real-time paddle conditions with ML-powered safety rating
      description: |
        **Real-time paddling condition assessment with machine learning intelligence**
        
        This endpoint analyzes current weather, marine, and environmental conditions to provide an actionable paddle safety rating with detailed condition breakdown.
        
        ### 🧠 Machine Learning Features
        - **Production ML Model**: GradientBoostingRegressor trained on 50,000+ condition combinations
        - **Marine Data Integration**: Real-time wave height and water temperature when available
        - **Smart Penalty System**: Transparent adjustments for extreme conditions
        - **0.5 Precision Ratings**: Only 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0 ratings
        
        ### 📊 Response Content
        - Current weather conditions with marine data
        - ML-generated paddle rating with confidence level  
        - Detailed penalty breakdown and justification
        - Safety warnings and actionable recommendations
        - Beaufort wind scale analysis
        
        ### ⚡ Performance
        **~250ms average response time** - Optimized for real-time mobile applications
        
        ### 🔄 Consistency Guarantee
        Identical conditions will produce identical ratings across `paddleScore` and `fastForecast` endpoints
        
      parameters:
        - name: lat
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: -90.0
            maximum: 90.0
            example: 32.881187
          description: |
            **Latitude coordinate in decimal degrees**
            
            Use with `lng` parameter for coordinate-based lookup. Use precise coordinates from the `paddlingOut` endpoint to ensure marine data accuracy.
        - name: lng
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: -180.0
            maximum: 180.0
            example: -96.929937
          description: |
            **Longitude coordinate in decimal degrees**
            
            Use with `lat` parameter for coordinate-based lookup. Use precise coordinates from the `paddlingOut` endpoint to ensure marine data accuracy.
        - name: location
          in: query
          required: false
          schema:
            type: string
            pattern: '^-?\d+(\.\d+)?,-?\d+(\.\d+)?$'
            example: "32.881187,-96.929937"
          description: |
            **GPS coordinates in "latitude,longitude" format**
            
            Alternative to separate lat/lng parameters. Use precise coordinates from the `paddlingOut` endpoint to ensure marine data accuracy.
            
            **Format**: `lat,lng` (comma-separated, no spaces)
            **Example**: `32.881187,-96.929937`
        - name: spotId
          in: query
          required: false
          schema:
            type: string
            enum: ["trinity_river", "cottonwood_lake", "lake_crescent", "jenny_lake", "lake_mcdonald", "white_rock_lake", "lewisville_lake", "merrimack_river", "lake_union", "antero_reservoir", "lake_powell", "jackson_lake", "diablo_lake", "taylor_park", "kens_lake", "colorado_river", "ambazari_lake"]
            example: "trinity_river"
          description: |
            **Pre-configured location identifier for fastest lookup**
            
            Using `spotId` provides fastest response times (~150ms) as location data is pre-cached. Get valid `spotId` values from the `paddlingOut` endpoint.
            
            **Note**: Provide either coordinates (lat+lng OR location) OR spotId, but not both.
      responses:
        '200':
          description: Successfully retrieved current paddle conditions and rating
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaddleScoreResponse'
              examples:
                perfect_conditions:
                  summary: Perfect Conditions - Colorado High Altitude
                  description: Example of 5.0 rating with no penalties at high altitude location
                  value:
                    success: true
                    location:
                      name: "Buena Vista"
                      coordinates:
                        latitude: 38.781063
                        longitude: -106.277812
                      region: "Colorado"
                      country: "United States of America"
                    paddleScore:
                      rating: 5.0
                      interpretation: "Excellent - Perfect paddling conditions"
                      confidence: "high"
                      mlModelUsed: true
                      predictionSource: "ml-model"
                      modelType: "GradientBoostingRegressor"
                      originalRating: 5.0
                      penalties: []
                      totalPenalty: 0.0
                      roundedTo05Increments: true
                    conditions:
                      temperature: 75.2
                      windSpeed: 5.2
                      beaufortScale: 2
                      uvIndex: 6.2
                      visibility: 18.5
                      humidity: 42
                      cloudCover: 15
                      hasWarnings: false
                      waveHeight: null
                      waterTemp: null
                      marineDataAvailable: false
                    metadata:
                      source: "ml-model"
                      modelType: "GradientBoostingRegressor"
                      timestamp: "2025-08-23T15:45:22.120Z"
                      response_time_ms: 247
                
                extreme_heat_penalty:
                  summary: Extreme Heat Penalty - Texas Summer
                  description: Example showing -1.5 penalty for extreme heat conditions
                  value:
                    success: true
                    location:
                      name: "Location 32.881187, -96.929937"
                      coordinates:
                        latitude: 32.881187
                        longitude: -96.929937
                      region: "Texas"
                      country: "United States of America"
                    paddleScore:
                      rating: 2.0
                      interpretation: "Poor - Difficult conditions"
                      confidence: "high"
                      mlModelUsed: true
                      predictionSource: "ml-model"
                      modelType: "GradientBoostingRegressor"
                      originalRating: 3.5
                      penalties: ["Extreme heat (99.0°F): -1.5"]
                      totalPenalty: 1.5
                      roundedTo05Increments: true
                    conditions:
                      temperature: 99.0
                      windSpeed: 2.2
                      beaufortScale: 1
                      uvIndex: 9.8
                      visibility: 21.0
                      humidity: 33
                      cloudCover: 25
                      hasWarnings: true
                      waveHeight: 0.2
                      waterTemp: 31.4
                      marineDataAvailable: true
                    metadata:
                      source: "ml-model"
                      modelType: "GradientBoostingRegressor"
                      timestamp: "2025-08-23T15:45:22.120Z"
                      response_time_ms: 252
                
                spot_id_lookup:
                  summary: SpotId Lookup - Fast Cached Response
                  description: Example using spotId for fastest response
                  value:
                    success: true
                    location:
                      name: "Trinity River"
                      coordinates:
                        latitude: 32.881187
                        longitude: -96.929937
                      region: "Texas"
                      country: "United States of America"
                    paddleScore:
                      rating: 3.5
                      interpretation: "Good - Favorable conditions"
                      confidence: "high"
                      mlModelUsed: true
                      predictionSource: "ml-model"
                      modelType: "GradientBoostingRegressor"
                      originalRating: 4.0
                      penalties: ["High heat (87.0°F): -0.5"]
                      totalPenalty: 0.5
                      roundedTo05Increments: true
                    conditions:
                      temperature: 87.0
                      windSpeed: 8.2
                      beaufortScale: 3
                      uvIndex: 7.1
                      marineDataAvailable: true
                    metadata:
                      source: "ml-model"
                      response_time_ms: 158
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_location:
                  summary: Missing location parameters
                  value:
                    success: false
                    error: "Invalid input parameters"
                    details: ["Location required: provide coordinates (lat,lng) or spotId"]
                    usage:
                      endpoint: "/paddleScore"
                      parameters:
                        location_methods: [
                          "lat=32.881187&lng=-96.929937 (separate coordinates)",
                          "location=32.881187,-96.929937 (combined coordinates)",
                          "spotId=trinity_river (known spot ID)"
                        ]
                      examples:
                        coordinates: "/paddleScore?lat=32.881187&lng=-96.929937"
                        location_string: "/paddleScore?location=32.881187,-96.929937"
                        spot_id: "/paddleScore?spotId=trinity_river"
                invalid_coordinates:
                  summary: Invalid coordinate values
                  value:
                    success: false
                    error: "Invalid input parameters"
                    details: ["Latitude must be between -90 and 90"]
                    usage:
                      endpoint: "/paddleScore"
                      examples:
                        valid_coordinates: "/paddleScore?lat=32.881187&lng=-96.929937"
        '404':
          description: SpotId not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Paddling spot not found"
                spotId: "invalid_spot"
                available_via: "/paddlingOut"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        **Enterprise API Key Authentication**
        
        Contact api-support@kaayko.com for enterprise access and API key provisioning.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        **OAuth 2.0 Bearer Token Authentication**
        
        For service-to-service authentication and enterprise integrations.
...
  schemas:
    PaddlingLocation:
      type: object
      title: Paddling Location Details
      description: Complete information about a paddling location including coordinates, amenities, and media
      required:
        - id
        - title
        - location
      properties:
        id:
          type: string
          description: Unique location identifier for API references
          pattern: '^[a-z0-9_]+$'
          example: "trinity_river"
        title:
          type: string
          description: Human-readable location name
          minLength: 3
          maxLength: 100
          example: "Trinity River"
        subtitle:
          type: string
          description: Location context and geographical details
          maxLength: 200
          example: "Irving, Texas"
        text:
          type: string
          description: Paddling description, tips, and local insights
          maxLength: 500
          example: "Solitude, birds and a good workout with consistent flow year-round"
        location:
          type: object
          title: GPS Coordinates
          description: Precise GPS coordinates for navigation and API calls
          required:
            - latitude
            - longitude
          properties:
            latitude:
              type: number
              format: double
              minimum: -90.0
              maximum: 90.0
              description: Latitude in decimal degrees (WGS84)
              example: 32.881187
            longitude:
              type: number
              format: double  
              minimum: -180.0
              maximum: 180.0
              description: Longitude in decimal degrees (WGS84)
              example: -96.929937
        difficulty:
          type: string
          description: Paddling difficulty assessment
          enum: ["Easy", "Easy to Moderate", "Moderate", "Moderate to Advanced", "Advanced", "Expert"]
          example: "Easy to Moderate"
        elevation:
          type: integer
          description: Elevation above sea level in feet
          minimum: -100
          maximum: 15000
          example: 450
        features:
          type: array
          description: Key location features and characteristics
          items:
            type: string
          maxItems: 10
          example: ["Urban wildlife", "Year-round access", "Calm waters", "Scenic views"]
        parkingAvl:
          type: boolean
          description: Parking facilities available
          example: true
        restroomsAvl:
          type: boolean
          description: Restroom facilities available
          example: true
        permits_required:
          type: boolean
          description: Whether permits or fees are required
          default: false
          example: false
        youtubeURL:
          type: string
          format: uri
          description: Instructional or promotional YouTube video URL
          pattern: '^https://(www\.)?youtube\.com/watch\?v=.+$'
          example: "https://youtube.com/watch?v=trinity_river_guide"
        imgSrc:
          type: array
          description: High-resolution location images
          items:
            type: string
            format: uri
            description: Direct URL to location image
          maxItems: 5
          example: 
            - "https://kaayko.com/images/trinity_river_aerial.jpg"
            - "https://kaayko.com/images/trinity_river_launch.jpg"
        amenities:
          type: object
          description: Additional amenity details
          properties:
            picnicTables:
              type: boolean
              example: true
            grillsAvailable:
              type: boolean
              example: false
            shadedAreas:
              type: boolean
              example: true
            rentalEquipment:
              type: boolean
              example: false
...
    ErrorResponse:
      type: object
      title: Error Response
      description: Standardized error response format with detailed debugging information
      required:
        - success
        - error
        - code
        - timestamp
      properties:
        success:
          type: boolean
          description: Request success indicator (always false for errors)
          example: false
        error:
          type: string
          description: Human-readable error message
          maxLength: 200
          example: "Invalid coordinates"
        code:
          type: string
          description: Machine-readable error code for programmatic handling
          pattern: '^[A-Z_]+$'
          example: "INVALID_COORDINATES"
        details:
          type: string
          description: Additional technical details about the error
          maxLength: 500
          example: "Latitude must be between -90.0 and 90.0, longitude must be between -180.0 and 180.0"
        suggestion:
          type: string
          description: Helpful suggestion for resolving the error
          maxLength: 300
          example: "Use coordinates from paddlingOut API for accurate marine data"
        timestamp:
          type: string
          format: date-time
          description: UTC timestamp when error occurred
          example: "2025-08-23T15:45:22.120Z"
        requestId:
          type: string
          description: Unique identifier for this request (for support purposes)
          pattern: '^[a-f0-9-]+$'
          example: "req_8f9e2d1a-4b5c-6789-01ef-234567890abc"
        retryAfter:
          type: integer
          minimum: 0
          description: Seconds to wait before retrying (for rate limit errors)
          example: 60
        supportUrl:
          type: string
          format: uri
          description: URL for additional support or documentation
          example: "https://kaayko.com/docs/api/errors#invalid-coordinates"
...